+++ packages/name-controller/src/logger.ts
@@ -1,0 +1,5 @@
+import { createProjectLogger, createModuleLogger } from '@metamask/utils';
+
+export const projectLogger = createProjectLogger('name-controller');
+
+export { createModuleLogger };